vtt-sync: phase 4 extract array/section merge utilities
Move the board-state merge helpers (~300 lines) out of
board-interactions.js into a new utils/merge-helpers.js. All six
functions (cloneSectionSimple, cloneArraySimple,
mergeArrayByIdWithTimestamp, mergeSceneKeyedSection,
mergeSceneStatePreservingGrid, mergeBoardStateSnapshot) are pure and
had no module-level state, so the extraction is a straight move. The
three exported names are re-exported from board-interactions.js so
existing test imports keep working without churn.

Added utils/__tests__/merge-helpers.test.mjs with 18 unit tests
covering delta vs fullSync merges, grid preservation, combat
sequence-based newer-wins, fog revealedCells array coercion, and
metadata handling. Existing merge-helper coverage in
token-visibility-merge, fog-array-coercion, and board-state-poller
still passes unchanged. Behavior unchanged.

board-interactions.js drops from 19,787 to 19,492 lines.
