+++ dnd/vtt/assets/js/ui/board-interactions.js
@@ -37,6 +37,11 @@ import {
   broadcastStaminaSync,
   subscribeToStaminaSync,
 } from '../services/stamina-sync-service.js';
+import {
+  mergeArrayByIdWithTimestamp,
+  mergeSceneKeyedSection,
+  mergeBoardStateSnapshot,
+} from '../utils/merge-helpers.js';
 
 const OVERLAY_LAYER_PREFIX = 'overlay-layer-';
 let overlayLayerSeed = Date.now();
@@ -279,308 +284,14 @@ const TURN_FLASH_TONE_CLASSES = {
 const SHEET_SYNC_DEBOUNCE_MS = 400;
 const MALICE_VICTORIES_ACTION = 'fetch-victories';
 
-/**
- * Module-level merge functions for board state synchronization.
- * These are exported for testing and can be used independently of mountBoardInteractions.
- */
-
-function cloneSectionSimple(section) {
-  if (!section || typeof section !== 'object') {
-    return {};
-  }
-  try {
-    return JSON.parse(JSON.stringify(section));
-  } catch (error) {
-    return {};
-  }
-}
-
-function cloneArraySimple(arr) {
-  if (!Array.isArray(arr)) {
-    return [];
-  }
-  try {
-    return JSON.parse(JSON.stringify(arr));
-  } catch (error) {
-    return [];
-  }
-}
-
-/**
- * Merge arrays of entities by ID using timestamp-based conflict resolution.
- * Keeps the entry with the higher _lastModified timestamp.
- * @param {Array} existingArray - Current array of entities
- * @param {Array} incomingArray - Incoming array of entities
- * @returns {Array} Merged array with newer entries winning conflicts
- */
-export function mergeArrayByIdWithTimestamp(existingArray, incomingArray) {
-  const existing = Array.isArray(existingArray) ? existingArray : [];
-  const incoming = Array.isArray(incomingArray) ? incomingArray : [];
-
-  const byId = new Map(existing.map((item) => [item.id, item]));
-
-  incoming.forEach((item) => {
-    if (item && item.id) {
-      const existingItem = byId.get(item.id);
-      if (existingItem) {
-        // Compare timestamps - keep the newer one
-        const existingTime = existingItem._lastModified || 0;
-        const incomingTime = item._lastModified || 0;
-        if (incomingTime >= existingTime) {
-          byId.set(item.id, item);
-        }
-        // else: keep existing (it's newer)
-      } else {
-        // New item from server
-        byId.set(item.id, item);
-      }
-    }
-  });
-
-  return Array.from(byId.values());
-}
-
-/**
- * Merge scene-keyed objects (placements, templates, drawings) with timestamp-based resolution.
- * @param {Object} existingSection - Current scene-keyed object
- * @param {Object} incomingSection - Incoming scene-keyed object
- * @param {Object} options - Merge options
- * @param {boolean} options.fullSync - If true, incoming is authoritative (deletions propagate)
- * @returns {Object} Merged scene-keyed object
- */
-export function mergeSceneKeyedSection(existingSection, incomingSection, { fullSync = false } = {}) {
-  const existing = existingSection && typeof existingSection === 'object' ? existingSection : {};
-  const incoming = incomingSection && typeof incomingSection === 'object' ? incomingSection : {};
-
-  const merged = {};
-
-  // For full sync, only use scenes from incoming (deleted scenes should disappear)
-  // For delta, merge scenes from both existing and incoming
-  const allSceneIds = fullSync
-    ? new Set(Object.keys(incoming))
-    : new Set([...Object.keys(existing), ...Object.keys(incoming)]);
-
-  allSceneIds.forEach((sceneId) => {
-    const existingArray = existing[sceneId];
-    const incomingArray = incoming[sceneId];
-
-    if (incomingArray !== undefined) {
-      if (fullSync) {
-        // Full sync: incoming is authoritative for what items EXIST (deletion sync),
-        // but we still use timestamp-based merging for individual items to avoid
-        // reverting recent local changes (e.g. HP updates) with stale server data.
-        const incomingItems = Array.isArray(incomingArray) ? incomingArray : [];
-        const existingItems = Array.isArray(existingArray) ? existingArray : [];
-
-        if (existingItems.length === 0) {
-          try {
-            merged[sceneId] = JSON.parse(JSON.stringify(incomingArray));
-          } catch (e) {
-            merged[sceneId] = [];
-          }
-        } else {
-          // Build map of incoming items - these define what exists on the server
-          const incomingById = new Map();
-          incomingItems.forEach((item) => {
-            if (item && item.id) incomingById.set(item.id, item);
-          });
-          const existingById = new Map();
-          existingItems.forEach((item) => {
-            if (item && item.id) existingById.set(item.id, item);
-          });
-
-          // For each incoming item, keep the version with the newer _lastModified
-          const result = [];
-          incomingById.forEach((incomingItem, id) => {
-            const existingItem = existingById.get(id);
-            if (existingItem) {
-              const existingTime = existingItem._lastModified || 0;
-              const incomingTime = incomingItem._lastModified || 0;
-              result.push(incomingTime >= existingTime ? incomingItem : existingItem);
-            } else {
-              result.push(incomingItem);
-            }
-          });
-          // Items NOT in incoming are considered deleted (fullSync semantics)
-          try {
-            merged[sceneId] = JSON.parse(JSON.stringify(result));
-          } catch (e) {
-            merged[sceneId] = result;
-          }
-        }
-      } else {
-        // Delta: merge with existing, preserving local items not in incoming
-        merged[sceneId] = mergeArrayByIdWithTimestamp(existingArray, incomingArray);
-      }
-    } else if (existingArray !== undefined && !fullSync) {
-      // Only existing has data and we're not in full sync - keep it (clone to avoid mutation)
-      try {
-        merged[sceneId] = JSON.parse(JSON.stringify(existingArray));
-      } catch (e) {
-        merged[sceneId] = [];
-      }
-    }
-  });
-
-  return merged;
-}
-
-/**
- * Merge sceneState while preserving grid settings from existing state.
- * Grid size is a permanent scene setting that should NOT be overwritten by sync.
- * Only combat, overlay, and other transient state should be synced.
- * @param {Object} existingSceneState - Current scene state
- * @param {Object} incomingSceneState - Incoming scene state from server
- * @returns {Object} Merged scene state with preserved grid settings
- */
-function mergeSceneStatePreservingGrid(existingSceneState, incomingSceneState) {
-  const existing = existingSceneState && typeof existingSceneState === 'object' ? existingSceneState : {};
-  const incoming = incomingSceneState && typeof incomingSceneState === 'object' ? incomingSceneState : {};
-
-  const merged = {};
-  const allSceneIds = new Set([...Object.keys(existing), ...Object.keys(incoming)]);
-
-  allSceneIds.forEach((sceneId) => {
-    const existingEntry = existing[sceneId];
-    const incomingEntry = incoming[sceneId];
-
-    if (!incomingEntry || typeof incomingEntry !== 'object') {
-      // No incoming data for this scene - preserve existing if present
-      if (existingEntry && typeof existingEntry === 'object') {
-        try {
-          merged[sceneId] = JSON.parse(JSON.stringify(existingEntry));
-        } catch (e) {
-          merged[sceneId] = { grid: { size: 64, locked: false, visible: true } };
-        }
-      }
-      return;
-    }
-
-    // Clone incoming entry
-    let mergedEntry;
-    try {
-      mergedEntry = JSON.parse(JSON.stringify(incomingEntry));
-    } catch (e) {
-      mergedEntry = {};
-    }
-
-    // CRITICAL: Preserve grid settings from existing state
-    // Grid is a permanent scene setting, not transient board state
-    if (existingEntry && typeof existingEntry === 'object' && existingEntry.grid) {
-      mergedEntry.grid = JSON.parse(JSON.stringify(existingEntry.grid));
-    }
-
-    // CRITICAL: Protect combat state from stale server data.
-    // When a player's keepalive save overwrites the server's combat state
-    // with stale data, the next poll would bring that stale data back and
-    // revert the GM's changes ("popback" bug).  Compare sequence numbers
-    // (immune to clock drift) or updatedAt timestamps to keep the newer one.
-    if (existingEntry && typeof existingEntry === 'object' &&
-        existingEntry.combat && typeof existingEntry.combat === 'object' &&
-        mergedEntry.combat && typeof mergedEntry.combat === 'object') {
-      const existingSeq = Number(existingEntry.combat.sequence) || 0;
-      const incomingSeq = Number(mergedEntry.combat.sequence) || 0;
-      const existingTs = Number(existingEntry.combat.updatedAt) || 0;
-      const incomingTs = Number(mergedEntry.combat.updatedAt) || 0;
-
-      let keepExisting = false;
-      if (existingSeq > 0 && incomingSeq > 0) {
-        keepExisting = existingSeq > incomingSeq;
-      } else if (existingTs > 0 && incomingTs > 0) {
-        keepExisting = existingTs > incomingTs;
-      }
-
-      if (keepExisting) {
-        mergedEntry.combat = JSON.parse(JSON.stringify(existingEntry.combat));
-      }
-    }
-
-    // Merge fogOfWar: trust the incoming (server) state as authoritative.
-    // The server state is saved before Pusher broadcasts, so by the time
-    // the poller reads it the fog data reflects the GM's latest changes.
-    // Using a union here would prevent the GM from adding fog back
-    // (removing cells from revealedCells) because deleted cells would be
-    // restored from the player's existing local state.
-    if (existingEntry && typeof existingEntry === 'object' && existingEntry.fogOfWar &&
-        typeof existingEntry.fogOfWar === 'object') {
-      if (!mergedEntry.fogOfWar || typeof mergedEntry.fogOfWar !== 'object') {
-        // Incoming has no fogOfWar at all — use existing entirely.
-        mergedEntry.fogOfWar = JSON.parse(JSON.stringify(existingEntry.fogOfWar));
-      } else {
-        // Coerce arrays to plain objects (PHP encodes empty {} as [])
-        if (Array.isArray(mergedEntry.fogOfWar.revealedCells)) {
-          mergedEntry.fogOfWar.revealedCells = {};
-        }
-      }
-    }
-
-    // Final safety: ensure revealedCells is always a plain object after merge
-    if (mergedEntry.fogOfWar && Array.isArray(mergedEntry.fogOfWar.revealedCells)) {
-      mergedEntry.fogOfWar.revealedCells = {};
-    }
-
-    merged[sceneId] = mergedEntry;
-  });
-
-  return merged;
-}
-
-/**
- * Merge board state snapshots with timestamp-based conflict resolution for placements, templates, and drawings.
- * This ensures that concurrent updates don't cause data loss.
- * When incoming has _fullSync: true (from GET responses), deletions will propagate correctly.
- * IMPORTANT: Grid settings are NEVER overwritten by sync - they are permanent scene settings.
- * @param {Object} existing - Current board state
- * @param {Object} incoming - Incoming board state from server
- * @returns {Object} Merged board state
- */
-export function mergeBoardStateSnapshot(existing, incoming) {
-  if (!incoming || typeof incoming !== 'object') {
-    return existing ?? {};
-  }
-
-  // Check if this is a full sync (authoritative state from server GET)
-  // When fullSync is true, items not in incoming have been deleted
-  const fullSync = Boolean(incoming._fullSync);
-
-  // If no existing state, just use incoming
-  if (!existing || typeof existing !== 'object') {
-    return {
-      activeSceneId: typeof incoming.activeSceneId === 'string' ? incoming.activeSceneId : null,
-      mapUrl: typeof incoming.mapUrl === 'string' ? incoming.mapUrl : null,
-      placements: cloneSectionSimple(incoming.placements),
-      sceneState: cloneSectionSimple(incoming.sceneState),
-      templates: cloneSectionSimple(incoming.templates),
-      drawings: cloneSectionSimple(incoming.drawings),
-      overlay: cloneSectionSimple(incoming.overlay),
-      pings: cloneArraySimple(incoming.pings),
-      metadata: cloneSectionSimple(incoming.metadata ?? incoming.meta),
-    };
-  }
-
-  // Merge with timestamp-based conflict resolution for placements, templates, drawings
-  // Pass fullSync flag to enable deletion sync when receiving full state from server
-  const mergeOptions = { fullSync };
-  const snapshot = {
-    activeSceneId: typeof incoming.activeSceneId === 'string' ? incoming.activeSceneId : existing.activeSceneId,
-    mapUrl: typeof incoming.mapUrl === 'string' ? incoming.mapUrl : existing.mapUrl,
-    placements: mergeSceneKeyedSection(existing.placements, incoming.placements, mergeOptions),
-    // CRITICAL: Use mergeSceneStatePreservingGrid to preserve grid settings
-    // Grid is a permanent scene setting that should NOT be synced
-    sceneState: mergeSceneStatePreservingGrid(existing.sceneState, incoming.sceneState),
-    templates: mergeSceneKeyedSection(existing.templates, incoming.templates, mergeOptions),
-    drawings: mergeSceneKeyedSection(existing.drawings, incoming.drawings, mergeOptions),
-    overlay: cloneSectionSimple(incoming.overlay),
-    pings: cloneArraySimple(incoming.pings),
-  };
-
-  const metadata = cloneSectionSimple(incoming.metadata ?? incoming.meta);
-  if (metadata && typeof metadata === 'object' && Object.keys(metadata).length > 0) {
-    snapshot.metadata = metadata;
-  }
-
-  return snapshot;
-}
+// Board state merge helpers live in utils/merge-helpers.js (imported at
+// the top of this file). Re-export them so existing callers importing
+// these names from './board-interactions.js' keep working.
+export {
+  mergeArrayByIdWithTimestamp,
+  mergeSceneKeyedSection,
+  mergeBoardStateSnapshot,
+};
 
 export function mountBoardInteractions(store, routes = {}) {
   const board = document.getElementById('vtt-board-canvas');
